Include status and user reason on app detail.
