+++ dashboard/src/components/AppList/AppListItem.tsx
@@ -2,12 +2,8 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import Tooltip from "components/js/Tooltip";
-import {
-  InstalledPackageSummary,
-  InstalledPackageStatus_StatusReason,
-  installedPackageStatus_StatusReasonToJSON,
-} from "gen/kubeappsapis/core/packages/v1alpha1/packages";
-import { getPluginIcon } from "shared/utils";
+import { InstalledPackageSummary } from "gen/kubeappsapis/core/packages/v1alpha1/packages";
+import { getAppStatusLabel, getPluginIcon } from "shared/utils";
 import placeholder from "../../placeholder.png";
 import * as url from "../../shared/url";
 import InfoCard from "../InfoCard/InfoCard";
@@ -16,15 +12,6 @@ import "./AppListItem.css";
 export interface IAppListItemProps {
   app: InstalledPackageSummary;
   cluster: string;
-}
-
-function getAppStatusLabel(
-  statusReason: InstalledPackageStatus_StatusReason = InstalledPackageStatus_StatusReason.STATUS_REASON_UNSPECIFIED,
-): string {
-  // The JSON versions of the reasons are forced to follow the standard
-  // pattern STATUS_REASON_<reason> by buf.
-  const jsonReason = installedPackageStatus_StatusReasonToJSON(statusReason);
-  return jsonReason.replace("STATUS_REASON_", "").toLowerCase();
 }
 
 function AppListItem(props: IAppListItemProps) {
