+++ dashboard/src/components/AppView/PackageInfo/PackageInfo.tsx
@@ -4,8 +4,11 @@
 import {
   AvailablePackageDetail,
   InstalledPackageDetail,
+  InstalledPackageStatus_StatusReason,
 } from "gen/kubeappsapis/core/packages/v1alpha1/packages";
 import PackageUpdateInfo from "./PackageUpdateInfo";
+import { getAppStatusLabel } from "shared/utils";
+
 interface IPackageInfoProps {
   installedPackageDetail: InstalledPackageDetail;
   availablePackageDetail?: AvailablePackageDetail;
@@ -16,6 +19,17 @@ function PackageInfo({ installedPackageDetail, availablePackageDetail }: IPackag
     <section className="left-menu">
       {installedPackageDetail && (
         <>
+          {installedPackageDetail.status && (
+            <section className="left-menu-subsection" aria-labelledby="packageinfo-versions">
+              <div>
+                Status: <strong>{getAppStatusLabel(installedPackageDetail.status.reason)}</strong>
+              </div>
+              {installedPackageDetail.status.reason !==
+                InstalledPackageStatus_StatusReason.STATUS_REASON_INSTALLED && (
+                <div>{installedPackageDetail.status.userReason}</div>
+              )}
+            </section>
+          )}
           <section className="left-menu-subsection" aria-labelledby="packageinfo-versions">
             <h5 className="left-menu-subsection-title" id="packageinfo-versions">
               Versions
